Refactor backends to their separate modules
