+++ CMakeLists.txt
@@ -99,6 +99,8 @@ set(CF_SRCS
 	src/cute_draw.cpp
 	src/cute_image.cpp
 	src/cute_graphics.cpp
+	src/cute_graphics_sdlgpu.cpp
+	src/cute_graphics_gles.cpp
 	src/cute_aseprite_cache.cpp
 	src/cute_png_cache.cpp
 	src/cute_https.cpp
