+++ src/internal/cute_imgui_internal.h
@@ -18,6 +18,6 @@
 
 void cf_imgui_init();
 void cf_imgui_shutdown();
-void cf_imgui_draw(SDL_GPUTexture* swapchain_texture);
+void cf_imgui_draw();
 
 #endif // CF_IMGUI_INTERNAL_H
